requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.15.0
numpy
pydantic
Pillow
python-multipart